pyproject.toml
[project]
name = "gutters"
version = "0.1.0"
description = "GUTTERS - Guided Universal Transcendental Transformation & Evolutionary Response System"
authors = [{ name = "GUTTERS Project", email = "contact@gutters.dev" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    # Core Framework
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.6.1",
    "pydantic-settings>=2.0.3",
    "fastapi>=0.109.1",
    "uvicorn>=0.27.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "httptools>=0.6.1",
    "python-multipart>=0.0.9",
    "gunicorn>=23.0.0",

    # Database & Migrations
    "SQLAlchemy>=2.0.25",
    "SQLAlchemy-Utils>=0.41.1",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pgvector>=0.3.0",
    "greenlet>=2.0.2",

    # Authentication & Security
    "python-jose>=3.3.0",
    "bcrypt>=4.1.1",
    "pyjwt>=2.8.0",

    # UUID & Identifiers
    "uuid>=1.30",
    "uuid6>=2024.1.12",

    # HTTP & Async
    "httpx>=0.26.0",
    "aiohttp>=3.13.0",

    # Task Queue & Caching
    "redis>=5.0.1",
    "arq>=0.25.0",

    # CRUD & Admin
    "fastcrud>=0.19.2",
    "crudadmin>=0.4.2",

    # Code Quality
    "ruff>=0.11.13",
    "mypy>=1.16.0",

    # Logging & Display
    "structlog>=25.1.0,<=25.5.0",
    "rich>=14.2.0",

    # Astronomy & Astrology
    "timezonefinder>=6.2.0",
    "kerykeion==5.6.0",
    "pyswisseph==2.10.3.2",
    "geopy==2.4.1",
    "skyfield>=1.54",
    "pywebpush>=1.14.0",
    "croniter>=2.0.0",
    "numpy>=2.0.0",
    "scipy>=1.17.0",

    # LLM & AI
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "langchain-google-genai>=0.0.1",
    "openai>=1.60.0",
    "tenacity>=8.0.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-mock>=3.14.0",
    "faker>=26.0.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.hatch.build.targets.wheel]
include = ["src/"]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
exclude = [
    "work/",                      # Legacy/work copies and work-in-progress
    "tests/",                     # Test suite files
    "test_*.py",                  # Root-level test files
    "*_demo.py",                  # Demo/example scripts
    "*_test.py",                  # Root-level test files
    "scripts/",                   # Utility and migration scripts
    ".venv/",                     # Virtual environment
    "node_modules/",              # Frontend dependencies
    "build/",                     # Build artifacts
    "dist/",                      # Distribution artifacts
    ".git/",                       # Git metadata
    "__pycache__/",               # Python cache
    "*.egg-info/",                # Egg info
]

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741",  # Ambiguous variable name
    # "UP035", # deprecated-assertion
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.formparsers",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest-asyncio>=1.0.0",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
mypy_path = "src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "src.app.*"
disallow_untyped_defs = true
